app/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smarter-team"
version = "0.1.0"
description = "Multi-agent AI system for business automation"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Yasmine Seidu", email = "yasmine@smarterteam.ai"}
]
keywords = [
    "claude",
    "agent",
    "ai",
    "automation",
    "business",
    "multi-agent",
    "sdk"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    # Core Framework
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "python-multipart==0.0.18",
    # Task Queue & Caching
    "celery==5.4.0",
    "redis==5.2.1",
    # Database
    "sqlalchemy==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    # AI & SDK
    "anthropic==0.52.0",
    # HTTP & Async
    "httpx==0.28.1",
    "aiohttp==3.11.11",
    # Vector DB & Memory
    "pinecone==5.4.2",
    "zep-python==2.0.2",
    # Integrations
    "pydantic==2.10.4",
    "pydantic-settings==2.7.1",
    "python-dotenv==1.0.1",
    "google-auth>=2.25.0",
    "requests>=2.31.0",
    # Utilities
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "email-validator==2.2.0",
    "structlog==24.4.0",
    # Type hints
    "typing-extensions==4.12.2",
    "greenlet>=3.3.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "pytest-mock==3.14.0",

    # Code Quality
    "ruff==0.8.4",
    "mypy==1.14.1",
    "bandit==1.8.0",

    # Fast Type Checking (optional - 10-60x faster than mypy)
    "ty==0.0.1a8",

    # Security Scanning
    "semgrep==1.99.0",

    # Pre-commit
    "pre-commit==4.0.1",
    "detect-secrets==1.5.0",
]

test = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "pytest-mock==3.14.0",
    "httpx==0.28.1",
]

docs = [
    "mkdocs==1.6.1",
    "mkdocs-material==9.5.48",
]

[project.urls]
Repository = "https://github.com/yasmineseidu/smarter-team"
Documentation = "https://docs.smarterteam.ai"
Issues = "https://github.com/yasmineseidu/smarter-team/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["__tests__"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "celery.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pinecone.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "zep_cloud.*"
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".env",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "ARG002", # unused method argument (for overrides)
    "ARG001", # unused function argument
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
known-third-party = ["fastapi", "sqlalchemy", "celery"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"
